Fixed normal map calculation
Derived the orthonormal vector from gradients. c is the constant you
need to multiply the normal vector by to normalize it. So the
orthonormal vector is c<-DDX,-DDY,2>. Further computation is necessary
to set MipData because you need to map from vector space to color space.
The normal vector ranges between {-1,1} in x and y, and {0,1} in z.
These ranges need to be mapped to RGB which ranges between {0,255} for
each color component.