Grid size parameterization
Can now change grid size through blueprint. Re-derived scalar factor for
unit normal vector assuming gradient distance unit of 1 instead of 2.
Renamed simulate input parameter to TimeFactor because DeltaSeconds is
misleading. This parameter should be >=1 in order for waves to propagate
outside of the disturbance radius. Hooking this parameter up to
DeltaSeconds in blueprints would not work because this value would
always be less than 1.